Use persistent auth-token instead of user/pass (#2)
* Use persistent auth-token instead of user/pass

Before you had to save your plex username/password in cleartext for
this script to work. Now the config-flow is changed to create a
persistent token to authenticate with plex. Username/password is only
needed to create the token and doesn't need to be saved.

Updated README to be better readable and added instructions for Auth-Token.